requirements.txt
fastapi
granian[reload]
sqlmodel
asyncpg
pydantic
pydantic-ai
python-dotenv
requests
httpx
pytest
pytest-asyncio
greenlet
google-generativeai
